backend/requirements.txt
fastapi
uvicorn
pydantic==2.5.0
python-multipart==0.0.6
grpcio==1.60.0
grpcio-tools==1.60.0
websockets==12.0
aiofiles==23.2.1
python-dateutil==2.8.2